Refactor Plotter
- Delete unused FakePlotter
- Move plot routes into correct folder
