+++ src/workers_control/flask/routes/plots/__init__.py
@@ -1,0 +1,5 @@
+from . import routes
+
+__all__ = [
+    "routes",
+]
